txtDoubleLabel: Try to exclude wildcard TXT (#20)
As described in #20, the TXTDoubleLabel warning would appear when
a zone contained a wildcard TXT record (not awfully common).

We will suppress this warning when we detect that the response
was created due to a wildcard record.

So this zone should generate a warning:

*.example.org				IN	TXT	300 "abc"
_acme-challenge.example.org.example.org	IN	TXT	300 "xxx"

But this shouldn't:

*.example.org				IN	TXT	300 "abc"
_acme-challenge.example.org		IN	TXT	300 "xxx"

Not should this:

*.example.org				IN	TXT	300 "abc"