web/db.go: Delete tests older than 7 days
We currently keep tests indefinitely. This is not only unneccessarily expensive on storage, it is also privacy-unfriendly. Users probably do not expect us to keep results around forever.

Keep the tests for 7 days, because these test results are often used for debugging a problem and may be needed for reference until the problem is resolved. 7 days is a compromise to allow for some debugging without keeping data around forever.